Started fixing core migrations and seeding errors.

+++ database/migrations/2021_05_16_133537_sapphire_core.php
@@ -23,7 +23,7 @@ return new class extends Migration{
         Schema::create('domains', function(Blueprint $table){
             $table->increments('id');
             $table->string('domain')->unique();
-            $table->unsignedBigInteger('tenant_id');
+            $table->unsignedInteger('tenant_id');
             $table->foreign('tenant_id')->references('id')->on('tenants')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
